% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{active_metaflow_profile_app}
\alias{active_metaflow_profile_app}
\title{Create Shiny app for displaying the current active Metaflow profile}
\usage{
active_metaflow_profile_app(profile_df, profile, profile_path)
}
\arguments{
\item{profile_df}{A data frame containing profile information}

\item{profile}{The active profile}

\item{profile_path}{path to the profile file}
}
\value{
A Shiny app object
}
\description{
Create Shiny app for displaying the current active Metaflow profile
}
\keyword{internal}
