% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{as.character.s3_object}
\alias{as.character.s3_object}
\title{Convert s3_object to character}
\usage{
\method{as.character}{s3_object}(x, ...)
}
\arguments{
\item{x}{An s3_object}

\item{...}{Additional arguments passed to as.character}
}
\description{
Convert s3_object to character
}
