% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{as.data.frame.s3_object}
\alias{as.data.frame.s3_object}
\title{Convert s3_object to data.frame}
\usage{
\method{as.data.frame}{s3_object}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{An s3_object}

\item{row.names}{NULL or a character vector giving the row names for the data frame}

\item{optional}{A logical value. If TRUE, setting row names and converting column names is optional}

\item{...}{Additional arguments passed to as.data.frame}
}
\description{
Convert s3_object to data.frame
}
