% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_metaflow_profiles_app}
\alias{available_metaflow_profiles_app}
\title{Create Shiny app to display all Metaflow profiles}
\usage{
available_metaflow_profiles_app(profiles_df)
}
\arguments{
\item{profiles_df}{profiles df with 'Profile' and 'path'.}
}
\value{
A Shiny app object.
}
\description{
This function creates a Shiny app that displays all available Metaflow
profiles and allows the user to set the active profile.
}
\keyword{internal}
