% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_metaflow_profiles_server}
\alias{available_metaflow_profiles_server}
\title{Create server function for profiles app}
\usage{
available_metaflow_profiles_server(profiles_df)
}
\description{
Create server function for profiles app
}
\keyword{internal}
