% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_metaflow_profiles_ui}
\alias{available_metaflow_profiles_ui}
\title{Create UI for profiles app}
\usage{
available_metaflow_profiles_ui()
}
\description{
Create UI for profiles app
}
\keyword{internal}
