% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Object.R
\name{convert_bool}
\alias{convert_bool}
\title{Convert Python boolean to R boolean}
\usage{
convert_bool(value)
}
\arguments{
\item{value}{The value to convert}
}
\value{
A logical value
}
\description{
Convert Python boolean to R boolean
}
\keyword{internal}
