% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_action_table}
\alias{create_action_table}
\title{Create an Action Table with Custom Rendering}
\usage{
create_action_table(data, action_column_fn, cell_renderer_fn)
}
\description{
Create an Action Table with Custom Rendering
}
\keyword{internal}
