% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{create_s3_client}
\alias{create_s3_client}
\title{Create a new S3Client}
\usage{
create_s3_client(
  tmproot = ".",
  bucket = NULL,
  prefix = NULL,
  run = NULL,
  s3root = NULL
)
}
\description{
Create a new S3Client
}
