% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{create_s3_client_with_role}
\alias{create_s3_client_with_role}
\title{Create a new S3Client with a specified AWS role}
\usage{
create_s3_client_with_role(role, ...)
}
\description{
Create a new S3Client with a specified AWS role
}
