% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{create_s3_client_with_session}
\alias{create_s3_client_with_session}
\title{Create a new S3Client with specified session variables}
\usage{
create_s3_client_with_session(session_vars, ...)
}
\description{
Create a new S3Client with specified session variables
}
