% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_styled_page}
\alias{create_styled_page}
\title{Create a standard Shiny page with custom styles}
\usage{
create_styled_page(..., additional_css = "")
}
\description{
Create a standard Shiny page with custom styles
}
\keyword{internal}
