% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_generic_table}
\alias{generate_generic_table}
\title{Generate a generic table}
\usage{
generate_generic_table(data, action_column = NULL, cell_renderer = NULL)
}
\description{
Generate a generic table
}
\keyword{internal}
