% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{get_available_metaflow_profiles}
\alias{get_available_metaflow_profiles}
\title{Get Available Metaflow Profiles}
\usage{
get_available_metaflow_profiles(return_data = FALSE)
}
\arguments{
\item{return_data}{Logical value indicating whether to return the profiles data as a data frame.}
}
\value{
If \code{return_data} is \code{TRUE}, a data frame with the profiles is returned.
Otherwise, \code{NULL} is returned invisibly after displaying the profiles in the console.
}
\description{
Retrieve and display a list of all available Metaflow profiles.
}
\examples{
\dontrun{
# Display all available profiles in the console
get_available_metaflow_profiles()

# Get the profiles data as a data frame
profiles_data <- get_available_metaflow_profiles(return_data = TRUE)
}
}
\seealso{
\code{\link[=get_profiles]{get_profiles()}}
}
