% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_custom_styles}
\alias{get_custom_styles}
\title{Get custom CSS styles for Shiny apps}
\usage{
get_custom_styles(additional_css = "")
}
\description{
Get custom CSS styles for Shiny apps
}
\keyword{internal}
