% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_profile_descriptions}
\alias{get_profile_descriptions}
\title{Get Descriptions for Active Metaflow Profile Keys}
\usage{
get_profile_descriptions(profile)
}
\description{
Get Descriptions for Active Metaflow Profile Keys
}
\keyword{internal}
