% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{get_s3_client}
\alias{get_s3_client}
\title{Get S3 Client}
\usage{
get_s3_client(
  s3_role_arn = NULL,
  s3_session_vars = NULL,
  s3_client_params = NULL
)
}
\arguments{
\item{s3_role_arn}{Optional character string specifying the AWS role ARN.}

\item{s3_session_vars}{Optional list of session variables.}

\item{s3_client_params}{Optional list of client parameters.}
}
\value{
A list containing the S3 client and any error that occurred.
}
\description{
Get S3 Client
}
