% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_set_active_profile}
\alias{handle_set_active_profile}
\title{Handle Set Active Profile}
\usage{
handle_set_active_profile(profile_name, profile_path)
}
\description{
Handle Set Active Profile
}
\keyword{internal}
