% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_metaflow.R
\name{install_metaflow}
\alias{install_metaflow}
\title{Install Metaflow}
\usage{
install_metaflow(version = NULL, ..., restart_session = TRUE, from_git = FALSE)
}
\arguments{
\item{version}{for specific version of Metaflow, e.g. '2.7.9'}

\item{...}{other arguments passed to [reticulate::py_install()].}

\item{restart_session}{Restart R session after installing (note this will only occur within RStudio).}

\item{from_git}{install the recent GitHub version of Metaflow}
}
\description{
Install Metaflow python
}
