% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_interactive_viewer_available}
\alias{is_interactive_viewer_available}
\title{Check if Interactive Viewer is Available}
\usage{
is_interactive_viewer_available()
}
\description{
Check if Interactive Viewer is Available
}
\keyword{internal}
