% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{is_valid_metaflow_home}
\alias{is_valid_metaflow_home}
\title{Check if a provided directory exists and contains valid config files}
\usage{
is_valid_metaflow_home(dir_path)
}
\description{
Check if a provided directory exists and contains valid config files
}
\keyword{internal}
