% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{metaflow-package}
\alias{metaflow-package}
\alias{metaflow}
\title{metaflow - Metaflow for R}
\description{
The version number of the metaflow R package is synced
with the version number of the python version.
}
\details{
Check out `?metaflow` for a full list of the functions provided by this
package.
}
\section{Install the `metaflow` Python package}{


* [install_metaflow()]
}

