% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\docType{data}
\name{metaflow}
\alias{metaflow}
\title{metaflow module
User can access functions/modules in metaflow that are not exposed through the
exported R functions.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
metaflow
}
\description{
metaflow module
User can access functions/modules in metaflow that are not exposed through the
exported R functions.
}
\keyword{datasets}
