% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metaflow_profile_viewer}
\alias{metaflow_profile_viewer}
\title{Display Metaflow Profile Information}
\usage{
metaflow_profile_viewer(data_df, shiny_app_func, title, ...)
}
\description{
This function displays Metaflow profile information either via a Shiny app in the RStudio Viewer
or provides a warning if the interactive viewer is not available.
}
\keyword{internal}
