% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_metaflow.R
\name{metaflow_version}
\alias{metaflow_version}
\title{Get the installed Metaflow version}
\usage{
metaflow_version()
}
\value{
A character string representing the Metaflow version.
}
\description{
\code{metaflow_version()} returns the version of the installed Metaflow Python
package.
}
\examples{
metaflow_version()

}
