% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{mf}
\alias{mf}
\title{Handle to the \code{metaflow} module}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 0.
}
\usage{
mf
}
\value{
Module(metaflow)
}
\description{
Handle to the \code{metaflow} module
}
\keyword{internal}
