% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{print.s3_object}
\alias{print.s3_object}
\title{Print method for s3_object}
\usage{
\method{print}{s3_object}(x, ...)
}
\arguments{
\item{x}{An s3_object}

\item{...}{Additional arguments passed to print}
}
\description{
Print method for s3_object
}
