% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{process_metaflow_profile_data}
\alias{process_metaflow_profile_data}
\title{Helper function to process and display Metaflow profile data}
\usage{
process_metaflow_profile_data(
  data,
  title,
  return_data = FALSE,
  key_col = "Key",
  value_col = "Value"
)
}
\description{
Helper function to process and display Metaflow profile data
}
\keyword{internal}
