% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{profile_detail_server}
\alias{profile_detail_server}
\title{Profile Detail Module Server}
\usage{
profile_detail_server(id, profile, profile_path, profile_df)
}
\description{
Profile Detail Module Server
}
\keyword{internal}
