% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{read_profile_json}
\alias{read_profile_json}
\title{Read a Metaflow profile configuration file}
\usage{
read_profile_json(path)
}
\description{
Read a Metaflow profile configuration file
}
\keyword{internal}
