% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3.R
\name{S3}
\alias{S3}
\title{R6 Class Wrapper for Metaflow S3 Client}
\description{
R6 Class Wrapper for Metaflow S3 Client

R6 Class Wrapper for Metaflow S3 Client
}
\details{
This class provides an R interface to the Metaflow S3 client functionality.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_s3}}{Python S3 object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3-new}{\code{S3$new()}}
\item \href{#method-S3-get}{\code{S3$get()}}
\item \href{#method-S3-put}{\code{S3$put()}}
\item \href{#method-S3-get_many}{\code{S3$get_many()}}
\item \href{#method-S3-put_many}{\code{S3$put_many()}}
\item \href{#method-S3-list_paths}{\code{S3$list_paths()}}
\item \href{#method-S3-list_recursive}{\code{S3$list_recursive()}}
\item \href{#method-S3-info}{\code{S3$info()}}
\item \href{#method-S3-info_many}{\code{S3$info_many()}}
\item \href{#method-S3-get_all}{\code{S3$get_all()}}
\item \href{#method-S3-get_recursive}{\code{S3$get_recursive()}}
\item \href{#method-S3-put_files}{\code{S3$put_files()}}
\item \href{#method-S3-close}{\code{S3$close()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3-new}{}}}
\subsection{Method \code{new()}}{
Create a new S3Client object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to the Python S3 constructor}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-get"></a>}}
\if{latex}{\out{\hypertarget{method-S3-get}{}}}
\subsection{Method \code{get()}}{
Get a single object from S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$get(key, return_missing = FALSE, return_info = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Object key}

\item{\code{return_missing}}{If TRUE, return a missing object instead of raising
an exception}

\item{\code{return_info}}{If TRUE, fetch content-type and user metadata}

\item{\code{...}}{Additional arguments passed to the Python get method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An S3Object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-put"></a>}}
\if{latex}{\out{\hypertarget{method-S3-put}{}}}
\subsection{Method \code{put()}}{
Put a single object to S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$put(key, obj, overwrite = TRUE, content_type = NULL, metadata = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Object key}

\item{\code{obj}}{Object to store}

\item{\code{overwrite}}{If TRUE, overwrite existing object}

\item{\code{content_type}}{MIME type for the object}

\item{\code{metadata}}{A list of additional headers to be stored as metadata}

\item{\code{...}}{Additional arguments passed to the Python put method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
URL of the stored object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-get_many"></a>}}
\if{latex}{\out{\hypertarget{method-S3-get_many}{}}}
\subsection{Method \code{get_many()}}{
Get many objects from S3 in parallel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$get_many(keys, return_missing = FALSE, return_info = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{List of object keys}

\item{\code{return_missing}}{If TRUE, return missing objects instead of raising
exceptions}

\item{\code{return_info}}{If TRUE, fetch content-type and user metadata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-put_many"></a>}}
\if{latex}{\out{\hypertarget{method-S3-put_many}{}}}
\subsection{Method \code{put_many()}}{
Put many objects to S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$put_many(key_objs, overwrite = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key_objs}}{List of key-object pairs or S3PutObjects}

\item{\code{overwrite}}{If TRUE, overwrite existing objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of (key, url) pairs for uploaded objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-list_paths"></a>}}
\if{latex}{\out{\hypertarget{method-S3-list_paths}{}}}
\subsection{Method \code{list_paths()}}{
List the next level of paths in S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$list_paths(keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{List of paths to list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-list_recursive"></a>}}
\if{latex}{\out{\hypertarget{method-S3-list_recursive}{}}}
\subsection{Method \code{list_recursive()}}{
List all objects recursively under the given prefixes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$list_recursive(keys = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{List of prefixes to list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-info"></a>}}
\if{latex}{\out{\hypertarget{method-S3-info}{}}}
\subsection{Method \code{info()}}{
Get metadata about a single object in S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$info(key = NULL, return_missing = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{Object key}

\item{\code{return_missing}}{If TRUE, return a missing object instead of raising
an exception}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An S3Object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-info_many"></a>}}
\if{latex}{\out{\hypertarget{method-S3-info_many}{}}}
\subsection{Method \code{info_many()}}{
Get metadata about many objects in S3 in parallel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$info_many(keys, return_missing = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{List of object keys}

\item{\code{return_missing}}{If TRUE, return missing objects instead of raising
exceptions}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-get_all"></a>}}
\if{latex}{\out{\hypertarget{method-S3-get_all}{}}}
\subsection{Method \code{get_all()}}{
Get all objects under the prefix set in the S3 constructor
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$get_all(return_info = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{return_info}}{If TRUE, fetch content-type and user metadata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-get_recursive"></a>}}
\if{latex}{\out{\hypertarget{method-S3-get_recursive}{}}}
\subsection{Method \code{get_recursive()}}{
Get many objects from S3 recursively in parallel
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$get_recursive(keys, return_info = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{List of prefixes to download recursively}

\item{\code{return_info}}{If TRUE, fetch content-type and user metadata}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of S3Object instances
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-put_files"></a>}}
\if{latex}{\out{\hypertarget{method-S3-put_files}{}}}
\subsection{Method \code{put_files()}}{
Upload many local files to S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$put_files(key_paths, overwrite = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key_paths}}{List of key-path pairs or S3PutObjects}

\item{\code{overwrite}}{If TRUE, overwrite existing objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of (key, url) pairs for uploaded files
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3-close"></a>}}
\if{latex}{\out{\hypertarget{method-S3-close}{}}}
\subsection{Method \code{close()}}{
Close the S3 client and delete temporary files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3$close()}\if{html}{\out{</div>}}
}

}
}
