% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{S3Client}
\alias{S3Client}
\title{S3Client Class}
\description{
An R6 class that manages the S3 client for AWS authentication and access.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Client-new}{\code{S3Client$new()}}
\item \href{#method-S3Client-client}{\code{S3Client$client()}}
\item \href{#method-S3Client-error}{\code{S3Client$error()}}
\item \href{#method-S3Client-set_role}{\code{S3Client$set_role()}}
\item \href{#method-S3Client-set_session_vars}{\code{S3Client$set_session_vars()}}
\item \href{#method-S3Client-set_client_params}{\code{S3Client$set_client_params()}}
\item \href{#method-S3Client-clone}{\code{S3Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-new}{}}}
\subsection{Method \code{new()}}{
Create a new S3Client object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$new(
  s3_role_arn = NULL,
  s3_session_vars = NULL,
  s3_client_params = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{s3_role_arn}}{Optional character string specifying the AWS role ARN.}

\item{\code{s3_session_vars}}{Optional list of session variables.}

\item{\code{s3_client_params}}{Optional list of client parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-client"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-client}{}}}
\subsection{Method \code{client()}}{
Get the S3 client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$client()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The S3 client object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-error"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-error}{}}}
\subsection{Method \code{error()}}{
Get the error associated with the S3 client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$error()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The error object, if any.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-set_role"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-set_role}{}}}
\subsection{Method \code{set_role()}}{
Set the AWS role for the S3 client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$set_role(role)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{role}}{Character string specifying the AWS role ARN.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-set_session_vars"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-set_session_vars}{}}}
\subsection{Method \code{set_session_vars()}}{
Set the session variables for the S3 client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$set_session_vars(session_vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session_vars}}{List of session variables.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-set_client_params"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-set_client_params}{}}}
\subsection{Method \code{set_client_params()}}{
Set the client parameters for the S3 client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$set_client_params(client_params)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{client_params}}{List of client parameters.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Client-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3Client-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
