% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{S3Object}
\alias{S3Object}
\title{S3Object Class}
\description{
Represents a path or an object in S3, with an optional local copy.
S3Objects are not instantiated directly, but they are returned
by many methods of the S3Client.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{exists}}{Logical indicating if this key corresponds to an object in S3.}

\item{\code{url}}{Character string of the S3 location of the object.}

\item{\code{prefix}}{Character string of the prefix requested that matches this object.}

\item{\code{key}}{Character string corresponding to the key given to the get call that produced this object.}

\item{\code{size}}{Integer size of the object in bytes (or NULL if the key doesn't correspond to an object in S3).}

\item{\code{has_info}}{Logical indicating if this S3Object contains additional metadata.}

\item{\code{metadata}}{List of user-defined metadata (or NULL if no metadata is defined).}

\item{\code{content_type}}{Character string of the content-type of the S3 object (or NULL if not defined).}

\item{\code{last_modified}}{Integer Unix timestamp of the last modified time.}

\item{\code{encryption}}{Character string of the server-side-encryption type (or NULL if not set).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{range_info}}{A list containing information about partial downloads, or NULL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Object-new}{\code{S3Object$new()}}
\item \href{#method-S3Object-print}{\code{S3Object$print()}}
\item \href{#method-S3Object-is_partially_downloaded}{\code{S3Object$is_partially_downloaded()}}
\item \href{#method-S3Object-clone}{\code{S3Object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new S3Object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$new(py_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{py_obj}}{A Python S3Object instance.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-print"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-print}{}}}
\subsection{Method \code{print()}}{
Print method for S3Object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments passed to print (unused).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-is_partially_downloaded"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-is_partially_downloaded}{}}}
\subsection{Method \code{is_partially_downloaded()}}{
Check if the object is partially downloaded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$is_partially_downloaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical indicating if the object is partially downloaded.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
