% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_profile_observers}
\alias{setup_profile_observers}
\title{Setup Profile Observers}
\usage{
setup_profile_observers(profiles_df, input, output, session)
}
\description{
Setup Profile Observers
}
\keyword{internal}
