% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{show_active_metaflow_profile}
\alias{show_active_metaflow_profile}
\title{Show Active Metaflow Profile}
\usage{
show_active_metaflow_profile()
}
\value{
The function returns the result of \code{display_data_generic()}, which launches a Shiny app
to display the active profile details in the RStudio Viewer pane or an external browser.
}
\description{
Display information about the currently active Metaflow profile in a Shiny app viewer.
}
\examples{
\dontrun{
# Display the active profile in a Shiny app
show_active_metaflow_profile()
}
}
\seealso{
\code{\link[=get_active_profile]{get_active_profile()}}, \code{\link[=process_metaflow_profile_data]{process_metaflow_profile_data()}}
}
