% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{show_available_metaflow_profiles}
\alias{show_available_metaflow_profiles}
\title{Show Available Metaflow Profiles}
\usage{
show_available_metaflow_profiles()
}
\value{
The function returns the result of \code{display_data_generic()}, which launches a Shiny app
to display the list of available profiles in the RStudio Viewer pane or an external browser.
}
\description{
Display a list of all available Metaflow profiles in a Shiny app viewer.
}
\examples{
\dontrun{
# Display all available profiles in a Shiny app
show_available_metaflow_profiles()
}
}
\seealso{
\code{\link[=get_profiles]{get_profiles()}}, \code{\link[=process_metaflow_profile_data]{process_metaflow_profile_data()}}
}
