% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_profile_modal}
\alias{show_profile_modal}
\title{Show Profile Detail Modal}
\usage{
show_profile_modal(profile_id, profile, profile_path, profile_df)
}
\description{
Show Profile Detail Modal
}
\keyword{internal}
