% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{str.s3_object}
\alias{str.s3_object}
\title{Str method for s3_object}
\usage{
\method{str}{s3_object}(object, ...)
}
\arguments{
\item{object}{An s3_object}

\item{...}{Additional arguments passed to str}
}
\description{
Str method for s3_object
}
