% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{summary.s3_object}
\alias{summary.s3_object}
\title{Summary method for s3_object}
\usage{
\method{summary}{s3_object}(object, ...)
}
\arguments{
\item{object}{An s3_object}

\item{...}{Additional arguments passed to summary}
}
\description{
Summary method for s3_object
}
