% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3Object.R
\name{S3Object}
\alias{S3Object}
\title{R6 Class Wrapper for Metaflow S3Object}
\description{
R6 Class Wrapper for Metaflow S3Object

R6 Class Wrapper for Metaflow S3Object
}
\details{
This class represents an S3 object returned by the s3_client.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_obj}}{Python S3Object}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{url}}{The S3 URL of the object}

\item{\code{key}}{The S3 key of the object}

\item{\code{size}}{The size of the object in bytes}

\item{\code{path}}{The local path of the object if downloaded}

\item{\code{blob}}{The object's content as bytes}

\item{\code{text}}{The object's content as text}

\item{\code{metadata}}{The object's metadata}

\item{\code{content_type}}{The MIME type of the object}

\item{\code{downloaded}}{Whether the object has been downloaded}

\item{\code{exists}}{Whether the object exists in S3}

\item{\code{last_modified}}{The last modified timestamp of the object}

\item{\code{prefix}}{The S3 prefix of the object}

\item{\code{has_info}}{Whether the object has additional info}

\item{\code{encryption}}{The encryption status of the object}

\item{\code{range_info}}{Information about partial content requests}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Object-new}{\code{S3Object$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-new}{}}}
\subsection{Method \code{new()}}{
Create a new S3Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$new(py_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{py_obj}}{Python S3Object to wrap}
}
\if{html}{\out{</div>}}
}
}
}
