% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{S3Object}
\alias{S3Object}
\title{R6 Class Wrapper for Metaflow S3Object}
\description{
R6 Class Wrapper for Metaflow S3Object

R6 Class Wrapper for Metaflow S3Object
}
\details{
This class represents an S3 object returned by the s3_client.
}
\note{
This is a read-only property. It cannot be set directly.
}
\examples{

## ------------------------------------------------
## Method `S3Object$convert_bool`
## ------------------------------------------------

s3obj$convert_bool(TRUE)  # Returns TRUE
s3obj$convert_bool(get_py_false())  # Returns FALSE
s3obj$convert_bool(NULL)  # Returns FALSE
\dontrun{
s3obj$convert_bool("not a boolean")  # Throws an error
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{py_obj}}{Python S3Object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-S3Object-convert_bool}{\code{S3Object$convert_bool()}}
\item \href{#method-S3Object-new}{\code{S3Object$new()}}
\item \href{#method-S3Object-url}{\code{S3Object$url()}}
\item \href{#method-S3Object-text}{\code{S3Object$text()}}
\item \href{#method-S3Object-size}{\code{S3Object$size()}}
\item \href{#method-S3Object-exists}{\code{S3Object$exists()}}
\item \href{#method-S3Object-downloaded}{\code{S3Object$downloaded()}}
\item \href{#method-S3Object-path}{\code{S3Object$path()}}
\item \href{#method-S3Object-blob}{\code{S3Object$blob()}}
\item \href{#method-S3Object-content_type}{\code{S3Object$content_type()}}
\item \href{#method-S3Object-metadata}{\code{S3Object$metadata()}}
\item \href{#method-S3Object-last_modified}{\code{S3Object$last_modified()}}
\item \href{#method-S3Object-prefix}{\code{S3Object$prefix()}}
\item \href{#method-S3Object-key}{\code{S3Object$key()}}
\item \href{#method-S3Object-has_info}{\code{S3Object$has_info()}}
\item \href{#method-S3Object-encryption}{\code{S3Object$encryption()}}
\item \href{#method-S3Object-range_info}{\code{S3Object$range_info()}}
\item \href{#method-S3Object-clone}{\code{S3Object$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-convert_bool"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-convert_bool}{}}}
\subsection{Method \code{convert_bool()}}{
Convert a value to a boolean, handling both R logical and Python boolean objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$convert_bool(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The value to convert}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical value
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{s3obj$convert_bool(TRUE)  # Returns TRUE
s3obj$convert_bool(get_py_false())  # Returns FALSE
s3obj$convert_bool(NULL)  # Returns FALSE
\dontrun{
s3obj$convert_bool("not a boolean")  # Throws an error
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-new"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-new}{}}}
\subsection{Method \code{new()}}{
Create a new S3Object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$new(py_obj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{py_obj}}{Python S3Object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-url"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-url}{}}}
\subsection{Method \code{url()}}{
Get the URL of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$url()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
URL as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-text"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-text}{}}}
\subsection{Method \code{text()}}{
Get the content of the S3 object as text
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$text()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Content as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-size"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-size}{}}}
\subsection{Method \code{size()}}{
Get the size of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Size in bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-exists"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-exists}{}}}
\subsection{Method \code{exists()}}{
Check if the object exists in S3
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean indicating existence
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-downloaded"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-downloaded}{}}}
\subsection{Method \code{downloaded()}}{
Check if the object has been downloaded
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$downloaded()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean indicating if downloaded
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-path"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-path}{}}}
\subsection{Method \code{path()}}{
Get the local path of the downloaded object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$path()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Local path as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-blob"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-blob}{}}}
\subsection{Method \code{blob()}}{
Get the content of the S3 object as a byte string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$blob()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Content as raw bytes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-content_type"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-content_type}{}}}
\subsection{Method \code{content_type()}}{
Get the content type of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$content_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Content type as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-metadata"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-metadata}{}}}
\subsection{Method \code{metadata()}}{
Get the metadata of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Metadata as a list
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-last_modified"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-last_modified}{}}}
\subsection{Method \code{last_modified()}}{
Get the last modified timestamp of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$last_modified()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Last modified timestamp as a POSIXct object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-prefix"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-prefix}{}}}
\subsection{Method \code{prefix()}}{
Get the prefix requested that matches this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$prefix()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Prefix as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-key"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-key}{}}}
\subsection{Method \code{key()}}{
Get the key corresponding to the get call that produced this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$key()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Key as a string
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-has_info"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-has_info}{}}}
\subsection{Method \code{has_info()}}{
Check if this S3Object contains additional metadata
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$has_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Boolean indicating if additional metadata is available
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-encryption"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-encryption}{}}}
\subsection{Method \code{encryption()}}{
Get the encryption type of the S3 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$encryption()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Encryption type as a string or NULL if not defined
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-range_info"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-range_info}{}}}
\subsection{Method \code{range_info()}}{
Get information about a partially downloaded object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$range_info()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list with total_size, request_offset, and request_length, or NULL if not applicable
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-S3Object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-S3Object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{S3Object$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
